Create Spotify API skeleton and database tables
